Use platform specific directories for packaging gem files.
Building gems for distinct platforms within one rake run conflicted
in the lib/ directory, because all binary extensions were stored
there and were copied from there to become packaged.
This resulted in storing a binary extension of a wrong platform in
a gem.

To avoid this, the files to package are now stored in a platform specific
stage directory and are used to package the specific gem.