Fake alternate i386-mswin32-60 platform when updating config
This platform is binary compatible with i386-mingw32 and is used
to keep compatibility with VC6/MSVCRT.dll builds of Ruby.

`-60` refers to the MSVCRT version (6.0) used by Visual C 6.0 back
in the day, even that today real version is 7.x

Other builds of Ruby with Visual C uses a different version of
MSVCRT.dll (8.0, 9.0 10.0) and because of that, are not binary
compatible with this.