Restore compatibility with RubyGems platforms
In 0.9.0 the support for usage of x86-mingw32 platforms was removed for
cross-compile platforms.

In the wild, several gems use this syntax to indicate the platforms the
gem targets.

Since this is a massive breakage, a workaround need be put in place.

This change restores that functionality and accept also the
RUBY_PLATFORM one (i386-mingw32)