Change make program detection with `make' fallback
  Some platforms provide a working `make' program out of the box, but
some versions of this program won't return a successful exit status
code when given `-v' as argument.
  One such platform is FreeBSD, providing the `make' program in the base
system which is based on `pmake'. This version won't accept `-v':

$ make -v
usage: make [-BeikNnqrstWwX]
            [-C directory] [-D variable] [-d flags] [-f makefile]
            [-I directory] [-J private] [-j max_jobs] [-m directory] [-T file]
            [-V variable] [variable=value] [target ...]
zsh: exit 2     make -v

  This change ensures detection with `gmake -v` or `make -v` will
continue to work as expected, but adds a fallback on plain `make' when
tested successfully with the `command' shell builtin.