Update default ruby version for native tasks
When trying to build a native (non cross-compiled) gem I ran into #156.
To fix this, I updated the default for `ruby_versions` to `[ruby_ver]`
(which is set to `RUBY_VERSION`). This shouldn't affect cross-compiled
work since `ruby_versions` will be set to
`@ruby_versions_per_platform[platf]` when set.