Strip cross built shared library files while linking
The .so files of extensions are often manually stripped after compilation per task in a Rakefile.
This is done to reduce gem size and because debug information in release binaries isn't particular useful.
This manual step can be avoided, if the cross ruby version is built with stripping enabled.
Compatibility is ensured, since stripping files a second time doesn't make a difference.