Trigger compilation by 'cross' task and remove obsolete code
So far cross build of gems was triggered by 'rake cross native gem'.
The "cross" task adds the cross targets to the "native" task.
Invoking of "native" then compiles for all cross targets and for the native running ruby.
The "gem" task then packages everything into corresponding gem files.

This has the downside, that cross build always bundles native target and cross targets together.
If only cross targets are wanted, then they have to be triggered directly by "native:<target>" tasks.

Rake-compiler-dock-1.0 uses a separate image for each target and it leads to issues,
if the native target is triggered in addition to the image cross target.
This commit therefore changes "cross" to trigger the cross targets without the native running ruby.
The cross build of gems can now be done by:

    rake cross gem

Adding "native" to the command leads to the previous behaviour of building for cross and native rubies.

This commit also removes a bunch of code that predates to the logik before commit a3365c995770ceddba098937c2ec9a5e4b3914d9 .
To that time there wasn't a "stage" directory for cross rubies, so that the cross builds couldn't be proberly triggered.
The "Sooo brittle" code removed the native running Linux or MacOS target from extensions that are not intended to be cross compiled for the native running ruby, but it was no generic solution.
Getting rid of this code and trigger cross compile directly is far easier.