Call cross_compiling block for native platform too
Currently the cross_compiling block isn't called when cross compiling for the build platform.
This is due to the fact that define_native_tasks() is called first for the build platform
and later on for the cross platforms. The task "native:<gemname>:<platform>" is defined
only once and therefore only without the callback. The callback is therefore not called
for the cross platform that equals the build platform.

OMHO the root issue is that rake-compiler doesn't has a clear separation between
native and cross tasks. However I think calling cross_compiling for any binary gem
(native or cross) is a suitable workaround.

This fixes bug https://github.com/sparklemotion/nokogiri/issues/1991#issuecomment-592755871