handle nil in config options (broken in 1.2.2) (#217)
__Context__

In rake-compiler 1.2.1 and earlier, if an extension did something like
this, rake-compiler handled it gracefully:

```ruby
  Rake::ExtensionTask.new("gem", spec) do |ext|
    ext.config_options << ENV["EXTOPTS"] # this is probably going to be nil
  end
```

It is OK to do this because in 1.2.1 and earlier, the command array was
executed as:

```
  sh cmd.join(' ')
```

However, this was changed in 1.2.2 to

```
  sh *cmd
```

which of course is preferable. However, this breaks extensions [like
nokogiri that do something silly like
above](https://github.com/sparklemotion/nokogiri/actions/runs/5109347425/jobs/9184057367).

__Details__

I've fixed Nokogiri (see
https://github.com/sparklemotion/nokogiri/pull/2894) but I do think that
there may be other gems impacted by this change.

This PR does a few things:

- extract a method `make_makefile_cmd` so that we can unit test it
- backfill basic test coverage
- backfill test for #215
- write a test for this `nil` case, and fix it by returning the result
of `cmd.compact`

This is a big PR because of the extraction and test coverage. Please let
me know if you would prefer this to be done a different way.