Use ENV variables to set conf paths (#222)
closes: GH-202

Instead of creating `.rake-compiler-siteconf.rb` to override
`RbConfig::MAKEFILE_CONFIG` and `RbConfig::CONFIG`, set environment
variables to `make install` command. Since `RbConfig` is suppose to be
readonly, we shouldn't overwrite its values.

Additionally, this update addresses a bug caused by
`.rake-compiler-siteconf.rb`. The file's hardcoded destination path
causes `rake-compiler` to consistently create files in that location,
even if the project is renamed or moved. Consequently, changes to the
gem project's path do not trigger the file task to update the file. This
fix resolves this issue.

---------

Co-authored-by: Sutou Kouhei <kou@clear-code.com>