Don't generate additional rake tasks when files don't exist
If files is retrieved from Git and the task is executed after the
files are deleted but before they are committed to Git, an error
occurs because the files do not exist.

Getting files from Git is a common pattern because it's default of
newgems by generated by Bundler.
https://github.com/rubygems/rubygems/blob/ad394cadc73f2239b907e6cd02a0dfae2ef65999/bundler/lib/bundler/templates/newgem/newgem.gemspec.tt#L31-L35

I would like to avoid an error in this situation, so I fixed not
to create rake tasks for compile if files don't exist.