javaextensiontask.rb - check `javac` path using `JAVA_HOME` (#247)
JRuby-head is using JDK 21. This is not the default on GHA. `setup-ruby`
added code to change `JAVA_HOME`, assuming it would trigger the correct
JDK being used.

Puma tests against jruby-head, and it was failing when rake compile ran.
After reviewing the code, `javac` is being called without a path.

Added code to check if `File.join(ENV["JAVA_HOME"], "bin", "javac")`
exists. If it does, use the full path, otherwise fallback to current
behavior.

I changed the Gemfile in Puma to use the PR branch, and the job
correctly compiles.